% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{cross_category}
\alias{cross_category}
\title{Plot the cross-category effect matrix}
\usage{
cross_category(object)
}
\arguments{
\item{object}{A FAIRforecast object obtained with \code{\link{FAIR_train}}.}
}
\description{
Plots the matrix of the interaction strength. It averages the absolute value of the
cross-category marketing variables of the x-axis category in the model for the y-axis
category.
}
