% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\docType{package}
\name{FAIR}
\alias{FAIR}
\alias{FAIR-package}
\title{FAIR: A package for FAIR forecast}
\description{
The main functionality of this package is to implement the FAIR
method proposed by Gür Ali and Gürlek (2019) to any panel data of
store-category-level sales time series of a retail. It utilizes
orthogonalization, regularization via Elastic net, and extrapolation
to account for seasonality, focal- and cross-category marketing activity
as well as random disturbances to base sales. It can handle large
dimensionality and deal with regularization induced confounding.
}
\details{
Package:  FAIR

Type:     Package

Version:  0.1.0

Release Date: 09-19-19

License:  GPL-2
}
\references{
Gür Ali, Ö. and Gürlek, R. (2019) Automatic Interpretable Retail
Forecasting with Promotional Scenarios

Gür Ali, Ö. and Pınar, E. (2016). Multi-period-ahead forecasting
  with residual extrapolation and information sharing - Utilizing a multitude
  of retail series. International Journal of Forecasting, 32(2):502–517.

Hahn, P. R., Carvalho, C. M., Puelz, D., and He, J. (2018).
Regularization and confounding in linear regression for treatment effect
estimation. Bayesian Analysis, 13(1):163– 182.

Zou, H. and Hastie, T. (2005), Regularization and variable
selection via the elastic net. Journal of the Royal Statistical Society:
Series B (Statistical Methodology), 67: 301-320.
}
\author{
Özden Gür Ali and Ragıp Gürlek

Maintainer: Ragıp Gürlek \email{rgurlek@yahoo.com.tr}
}
