% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_predict.R
\name{FAIR_predict}
\alias{FAIR_predict}
\title{Predict sales for a new data}
\usage{
FAIR_predict(object, new_data, parallel = F)
}
\arguments{
\item{object}{A FAIRforecast object obtained with \code{\link{FAIR_train}}.}

\item{new_data}{A dataframe object that contains observations with
the following variables:
store identifier (factor), category identifier (factor), time id (numeric),
seasonality and calendar variables (factor or numeric),
and marketing variables (numeric).}

\item{parallel}{Logical. Should parallel computing be used? The default is
\code{F}}
}
\value{
A data.frame of the final forecast and its components for each
 \code{store}-\code{category}-\code{time_id} combination in \code{new_data}.
 The columns are:
 \item{Base_Sales}{The forecast based on established
 seasonality and trend in sales and marketing. Obtained with Stage 1 and
 adjusted for bias.}
 \item{Marketing_deviation_multiplier}{Multiplier based on the marketing
 scenario deviation from the established pattern.}
 \item{Disturbance_multiplier}{Multiplier reflecting the changes in base
 sales.}
 \item{Forecast}{The final forecast obtained by multiplying the components
 and subtracting 1. If more complex parts of the forecast (multipliers) are
 set to NA, they are ignored.}
}
\description{
Predict sales for a new data.
}
\details{
All components of the data is supplied in one data.frame and the
 column names should correspond to the ones used to obtain \code{object}.

\code{FAIR_predict} can forecast for periods which are after the
 last period of the training data and are in the forecasting horizon.
}
\references{
Gür Ali, Ö. and Gürlek, R. (2020) Automatic Interpretable Retail
Forecasting (FAIR) with Promotional Scenarios. International Journal of
Forecasting, forthcoming.
}
\seealso{
\code{\link{FAIR_train}}
}
