% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_train.R
\name{FAIR_train}
\alias{FAIR_train}
\title{Fit a FAIR model}
\usage{
FAIR_train(
  train_data,
  store,
  category,
  time_id,
  seasonality,
  marketing,
  cc_marketing = NULL,
  sales,
  lag = 2,
  K = 6,
  alpha = c(0, 0.1, 0.3, 0.5, 0.7, 0.9, 1),
  frequency = 365.25/7,
  horizon = 1,
  parallel = F
)
}
\arguments{
\item{train_data}{a dataframe object that contains observations with the following
variables: store identifier (factor), category identifier (factor), time id
(numeric), seasonality and calendar variables (factor or numeric), marketing
variables (numeric), and sales (numeric).}

\item{store}{a string. The name of the store identifier
column.}

\item{category}{a string. The name of the category
identifier column.}

\item{time_id}{a string. The name of the time
identifier column.}

\item{seasonality}{a character vector. The names of the seasonality and
calendar columns.}

\item{marketing}{a character vector. The names of the marketing columns.
The corresponding columns are expected to represent continuous variables and
to be non-negative.}

\item{cc_marketing}{a character vector. The names of the
cross-category marketing columns. Cross-category marketing variables are a
subset of the marketing variables that are assumed to have a cross-series
impact. The default for this is \code{NULL}.}

\item{sales}{a string. The name of the sales column. The corresponding column
is expected to be non-negative.}

\item{lag}{numeric. Number of lags of the marketing variables to consider.
The default is \code{2}}

\item{K}{numeric. Number of folds in CV. The default is \code{6}}

\item{alpha}{numeric vector of alpha values. These are used as alpha parameter
of \code{\link{glmnet}} function employed in Stage 2. The best value is
chosen with a CV. All values should be in [0,1]. The default is
\code{c(0, 0.1, 0.3, 0.5, 0.7, 0.9, 1)}}

\item{frequency}{numeric. Frequency of the timeseries. The default is
\code{365.25 / 7}}

\item{horizon}{numeric. Leadtime of the forecast. The default is \code{1}}

\item{parallel}{Logical. Should parallel computing be used? The default is
\code{F}}
}
\value{
The function outputs a list with the following objects:
 \item{fit}{A data.frame of the sales estimate and its components for each
 \code{store}-\code{category}-\code{time_id} combination in \code{train_data}.
 See the value section of \code{\link{FAIR_predict}} for details.}
 \item{models}{the model for each stage and a data.frame for in-sample bias}
 \item{pars}{parameters used for the training}
 \item{lag_data}{data.frame of the last periods, used by
 \code{FAIR_predict}}
}
\description{
\code{FAIR_train} fits a FAIR model to the provided data.
}
\details{
This function fits a FAIR model to forecast store-category-level
 retail sales time series. The data is assumed to be of panel nature where
 the series in the same store affects each other.

All components of \code{train_data} is supplied in one
 data.frame. The user specifies the names of the individual components with
 the corresponding arguments.

The series do not have to be of the same size and they may have gaps,
 i.e., the input data may lack rows for some
 \code{time_id}. The algorithm can work with missing values. However, if the proportion
 of the missing values in a series exceeds a certain threshold the series is
 ignored and the user is informed.
}
\references{
Gür Ali, Ö. and Gürlek, R. (2020) Automatic Interpretable Retail
Forecasting (FAIR) with Promotional Scenarios. International Journal of
Forecasting, forthcoming.
}
\seealso{
\code{\link{FAIR_predict}}
}
\author{
Özden Gür Ali and Ragıp Gürlek

Maintainer: Ragıp Gürlek \email{rgurlek@yahoo.com.tr}
}
