% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\docType{package}
\name{FAIRforecast}
\alias{FAIRforecast}
\title{FAIRforecast: A package for FAIR forecast}
\description{
FAIR is a forecasting tool developed to support decision making in a retail
environment. It provides multi-step-ahead sales forecasts at the category-store
level, which are based on an interpretable and transparent model. These aspects
make it an objective tool with which different promotional strategies
(scenarios) can be compared and insights can be generated. Additionally, the FAIRforecast
package provides plotting functions that generate figures showing the relative
strength of the interactions between categories and important promotional
variables. For more information, see \code{vignette("FAIRforecast")} and
Gür Ali & Gürlek (2020).
}
\details{
Package:  FAIRforecast

Type:     Package

Version:  0.1.0

Release Date: 03-15-20

License:  GPL-2
}
\references{
Gür Ali, Ö. and Gürlek, R. (2020) Automatic Interpretable Retail
Forecasting (FAIR) with Promotional Scenarios. International Journal of
Forecasting, forthcoming.

Gür Ali, Ö. and Pınar, E. (2016). Multi-period-ahead forecasting
  with residual extrapolation and information sharing - Utilizing a multitude
  of retail series. International Journal of Forecasting, 32(2):502–517.

Hahn, P. R., Carvalho, C. M., Puelz, D., and He, J. (2018).
Regularization and confounding in linear regression for treatment effect
estimation. Bayesian Analysis, 13(1):163– 182.

Zou, H. and Hastie, T. (2005), Regularization and variable
selection via the elastic net. Journal of the Royal Statistical Society:
Series B (Statistical Methodology), 67: 301-320.
}
\author{
Özden Gür Ali and Ragıp Gürlek

Maintainer: Ragıp Gürlek \email{rgurlek@emory.edu}

\url{https://github.com/rgurlek/FAIRforecast}
}
