% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{A sample retail sales dataset}
\format{A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}

}
Note that distributional variables do not add up to 1 because they exclude
non-discounted SKUs.}
\usage{
sample_data
}
\description{
A dataset containing the weekly category-store-level sales of a retail chain.
It provides the promotional activity as well as the seasoanlity and calendar
variables.
}
\keyword{datasets}
