% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{variable_importance}
\alias{variable_importance}
\title{Plot variable importance}
\usage{
variable_importance(object, n_vars = 10, categories = NULL, only_own = T)
}
\arguments{
\item{object}{A FAIRforecast object obtained with \code{\link{FAIR_train}}.}

\item{n_vars}{Number of top marketing variables to be included. The
default is 10.}

\item{categories}{Character vector of categories over which the averages
are calculated. The default, \code{NULL}, includes the all.}

\item{only_own}{If \code{TRUE}, only the focal variables are plotted.}
}
\description{
Plots the average absolute coefficients of the top marketing variables.
}
